pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.51.0",
    "mcp[cli]>=1.8.0",
    "prompt-toolkit>=3.0.51",
    "python-dotenv>=1.1.0",
    "openai>=1.0.0",
    "PyPDF2>=3.0.0",
    "python-docx>=0.8.11",
    "Pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "python-magic>=0.4.27",
    "chardet>=5.0.0",
]
